Fixed a bug in renaming GADTs.
So far, free variables in a constructor declaration has been treated as
unbound, but this commit has fixed the bug.

There still is a bug in GADT-style constructor declarations. For

   data T s where
     MkT : s -o T (S s)

is currently problematic, because it generates a constructor of type

   MkT : forall s. s ~ S s => s -o T (S s)

which is clearly undesirable. Fortunately, there is an easy workaround
for this issue: just avoid using same names. For example, the following
declaration has no problem.

   data T ss where
      MkT : s -o T (S s)

From the declaration, we obtain a constructor of the following type.

   MkT : forall ss. ss ~ S s => s -o T (S s)